warehouse-api/requirements.txt
django==3.2.3
django-extensions==3.1.3
ipython==7.24.0
flake8==3.9.2
flake8-django==1.1.2
flake8-builtins==1.5.3
flake8-import-order==0.18.1
flake8-print==4.0.0
faker==8.5.0
djangorestframework==3.12.4
psycopg2-binary==2.9.1
asgiref==3.8.1
backcall==0.2.0
decorator==5.1.1
Django==3.2.3
django-extensions==3.1.3
djangorestframework==3.12.4
Faker==8.5.0
flake8==3.9.2
flake8-builtins==1.5.3
flake8-django==1.1.2
flake8-import-order==0.18.1
flake8-print==4.0.0
ipython==7.24.0
jedi==0.19.1
matplotlib-inline==0.1.7
mccabe==0.6.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
prompt-toolkit==3.0.43
psycopg2-binary==2.9.1
ptyprocess==0.7.0
pycodestyle==2.7.0
pyflakes==2.3.1
Pygments==2.17.2
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sqlparse==0.5.0
text-unidecode==1.3
traitlets==5.14.3
typing_extensions==4.11.0
wcwidth==0.2.13
amqp==5.2.0
asgiref==3.8.1
backcall==0.2.0
billiard==4.2.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
decorator==5.1.1
Django==3.2.3
django-extensions==3.1.3
djangorestframework==3.12.4
Faker==8.5.0
flake8==3.9.2
flake8-builtins==1.5.3
flake8-django==1.1.2
flake8-import-order==0.18.1
flake8-print==4.0.0
ipython==7.24.0
jedi==0.19.1
kombu==5.3.7
matplotlib-inline==0.1.7
mccabe==0.6.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
prompt-toolkit==3.0.43
psycopg2-binary==2.9.1
ptyprocess==0.7.0
pycodestyle==2.7.0
pyflakes==2.3.1
Pygments==2.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
sqlparse==0.5.0
text-unidecode==1.3
traitlets==5.14.3
typing_extensions==4.11.0
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
